Move WITH clause handling code into dedicated directory
