+++ src/with_clause/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(SOURCES
+    ${CMAKE_CURRENT_SOURCE_DIR}/compression_with_clause.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/create_table_with_clause.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/with_clause_parser.c)
+
+target_sources(${PROJECT_NAME} PRIVATE ${SOURCES})
